Fixing build errors in x86 mode.

+++ util.c
@@ -226,7 +226,7 @@ void setWndPosDeferred(const SetWindowPosArgs *ops)
         goto fail;
     }
 
-    if (!(hdwp = BeginDeferWindowPos(count)))
+    if (!(hdwp = BeginDeferWindowPos((int) count)))
     {
         /* TODO error */
         goto fail;
